Create solar_pannel_placement.py
This is the solution to the CCSCNE Programming Contest for the Spring 2017. The problem is as follows:

Your program will except multiple test cases. Each list consists of an energy readings and the object is to find the sum of the readings and a continuous subset that contains the largest energy reading.